Signal handler IDs are gulong, not int

+++ libmatekbd/matekbd-desktop-config.h
@@ -43,7 +43,7 @@ struct _MatekbdDesktopConfig {
 
 	/* private, transient */
 	GSettings *settings;
-	int config_listener_id;
+	gulong config_listener_id;
 	XklEngine *engine;
 };
